Refactor Mac platform structure and update imports.
Reorganized Mac-specific packages, moving from "avplayer" to "compose" and "swing" to improve clarity and consistency. Updated relevant imports across the project to align with the new structure. This change aims to streamline navigation and maintain a consistent naming convention.
